TestGetProposalsFiltered migrate to simapp
